requirements.txt
Flask==2.2.2
pandas==2.1.1
Pillow==9.1.0
requests==2.31.0
google-generativeai==0.1.0
openpyxl==3.0.10
